Update 2.0.0
In this update the user is now able to use the program I called Beastsy which is a litteral pronounciation of the letter combo B.C.T.C. which stands for Bitmap Converter To C.

Anyway, you will now be able to get your .bmp file and convert it automatically to an .h (header) file.

If you have a very small image the canvas will auto size it to make it better visible, is your .bmp bigger than 500 x 500 pixels than you'll be allowed to pan over your image.